requirements.txt
# Data Analysis & Manipulation
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & Stats
scikit-learn>=1.3.0
lightgbm
category_encoders
imbalanced-learn
statsmodels
gensim
nltk

# Visualization
matplotlib
seaborn
wordcloud

# Utilities
tqdm
pyyaml
python-dateutil

# Jupyter Support
ipykernel